+++ tomotok/core/io/diag.py
@@ -4,12 +4,12 @@
 """
 Template class for general tomographic diagnostic system.
 """
+import json
 import warnings
-
-import h5py
-import numpy as np
-
 from pathlib import Path
+from typing import Tuple, Union
+
+import numpy as np
 
 
 class Dsystem(object):
@@ -21,136 +21,44 @@ class Dsystem(object):
     chnl_index : dict
         dict maping channel to detector
     geometry_path : pathlib.Path
-        path to location with saved geometrical data
+        path to location containing diagnostics system geometry data
     name : str
         name of diagnostic system
     n_dets : int
         number of cameras with separate calibration files, (to be removed)
     """
 
-    def __init__(self, shot, geometry_path='', **kw):
+    def __init__(self, shot, geometry_path: Union[str, Path] = '', **kw):
+        """
+        Parameters
+        ----------
+        shot : int
+            shot number
+        geometry_path : str or pathlib.Path, optional
+            path to geometry folder, by default '' (current folder)
+        """
         super().__init__()
         self.name = self.__class__.__name__
         self.shot = shot
-        self.geometry_path = Path(geometry_path)
+        self.geometry_path = Path(geometry_path).expanduser()
         self.chnl_index = {}
-        self.calb_0 = None
-        self.n_dets = None
         self.keywords = kw
         return
 
-    @property
-    def data_name(self):
-        """
-        Template data file name
-        """
-        name = type(self).__name__
-        return '{nm}_{shot}'.format(nm=name, shot=self.shot)
-
-    def download_data(self, **kwargs):
+    def download_data(self, **kwargs) -> dict:
         """
         Downloads data from network/database using `source` method or database.
 
         Returns
         -------
         dict
+            holds time axis and data
         """
-        raise NotImplementedError("Download data not defined in subclass")
+        raise NotImplementedError("Download data should be defined in a subclass.")
 
-    def load_data(self, loc):
+    def load_los(self, loc: Union[str, Path] = None, check_data: bool = False) -> dict:
         """
-        Loads hdf file containing data.
-
-        If loc is path to a directory, default data file name from property data_name is assumed.
-
-        Parameters
-        ----------
-        loc : str or Path
-            Path to file with signal data. Can be full path or path to folder if default name is used.
-
-        Returns
-        -------
-        dict
-            keys values, chnl, time
-        """
-        loc = Path(loc)
-        data = dict()
-        if loc.is_dir():
-            loc = Path(loc) / self.data_name
-        with h5py.File(str(loc), 'r') as array:
-            for key in array:
-                data[key] = array[key][...]
-        return data
-
-    # TODO integrate into fetch and remove?
-    @staticmethod
-    def save_data(loc, data):
-        """
-        Saves downloaded detector data to data_path
-
-        Parameters
-        ----------
-        loc : str or pathlib.Path
-            saving path with file name
-        data : dict
-            contains signal values, channel numbers and time stamps
-        """
-        loc = Path(loc)
-        try:
-            loc.parent.mkdir(parents=True, exist_ok=True)
-            with h5py.File(str(loc), 'w') as hdf:
-                for key in ['values', 'chnl', 'time']:
-                    hdf[key] = data[key]
-        except OSError as e:
-            warnings.warn('Detector data were not saved: {}'.format(e))
-        return
-
-    def fetch_data(self, data_path='', source=None, save=False, force_dl=False):
-        """
-        Loads data from data storage specified by data_path or downloads data
-        from network storage.
-
-        Parameters
-        ----------
-        data_path : str or Path, optional
-            default is working directory
-        source : :obj:`string`, optional
-            specifies method for data download
-        save : bool, optional
-            switches saving to local storage
-        force_dl : bool, optional
-            forces download of data
-
-        Returns
-        -------
-        dict
-        """
-        floc = Path(data_path)
-        try:
-            assert force_dl is False
-            data = self.load_data(floc)
-        except (OSError, AssertionError):
-            data = self.download_data(source=source)
-            if save:
-                self.save_data(floc / self.data_name, data)
-        return data
-
-    def get_calb(self):
-        """
-        Loads preset calibrations for detectors, if not specified returns array of ones with length n_dets.
-        """
-        warnings.warn(
-            'Method get_calb not specified in Dsystem subclass. ' +
-            'Using default get_calb, which returns calb_0 or ones.',
-            Warning)
-        # TODO add method for loading detector calibrations
-        if self.calb_0 is None:
-            self.calb_0 = np.ones(self.n_dets)
-        return self.calb_0
-
-    def load_los(self, loc=None):
-        """
-        Loads line of sights geometry from hdf file.
+        Loads line of sights geometry from json file.
 
         Parameters
         ----------
@@ -160,21 +68,31 @@ class Dsystem(object):
         Returns
         -------
         dict
-            with coordinates of los start and end points with shape (2, #chords, 3)
+            a key for each camera that holds dictionary with `startpoints` and `endpoints` keys
+            each coordinates list has shape (#chords, 3)
         """
         los = {}
         if loc is None:
-            los_path = self.geometry_path / self.name
+            los_path = self.geometry_path / f'{self.name}_los.json'
         else:
-            los_path = Path(loc)
-        with h5py.File(los_path, 'r') as fl:
-            for key in fl.keys():
-                st = fl[key]['start'][:]
-                dir = fl[key]['dirs'][:]
-                los[key] = np.stack((st, st+dir), axis=0)
+            los_path = Path(loc).expanduser()
+        with open(los_path, 'r') as fl:
+            los = json.load(fl)
+        if not isinstance(los, dict):
+            raise ValueError('los file must contain dictionary')
+        if check_data:
+            for key in los:
+                if 'startpoints' not in los[key] or 'endpoints' not in los[key]:
+                    raise ValueError('Los file must contain startpoints and endpoints for each detector.')
+                if len(los[key]['startpoints']) != len(los[key]['endpoints']):
+                    raise ValueError('Start and end points lists must have same length and shape.')
+                sp_coords_fail = any([len(sp) != 3 for sp in los[key]['startpoints']])
+                ep_coords_fail = any([len(ep) != 3 for ep in los[key]['endpoints']])
+                if sp_coords_fail or ep_coords_fail:
+                    raise ValueError('Start and end points arrays must provide 3D Cartesian coordinates for each point.')
         return los
-    
-    def get_chord_geometry(self):
+
+    def get_chord_geometry(self) -> Tuple[np.ndarray, np.ndarray]:
         """
         Loads chord geometry from `geometry_path`.
 
@@ -183,12 +101,15 @@ class Dsystem(object):
         Each row represents one detector.
         Number of rows and columns should be same in all files containing coordinates
 
+        .. deprecated:: 2.0
+            use :meth:`load_los` instead
+
         Returns
         -------
         xchords, ychords : numpy.ndarray
             arrays with coordinates of chord points with shape (#chords, #points)
         """
-        warnings.warn('Support for LoS saving to txt will be removed. Transfer your coordinates to hdf or json.', FutureWarning)
+        warnings.warn('x/y chord format was deprecated. Use start/end points', DeprecationWarning)
         geometry_path = self.geometry_path
         ychords = np.zeros((0, 0))
         xchords = np.zeros((0, 0))
@@ -226,6 +147,9 @@ class Dsystem(object):
         Loads pinhole coordinates for each chord from `pinholes.txt` located
         in geometry folder.
 
+        .. deprecated:: 2.0
+            no clear universal use case
+
         Returns
         -------
         ph : numpy.ndarray
@@ -238,6 +162,9 @@ class Dsystem(object):
     def get_chord_widening(self):
         """
         Tries to load widening coefficients from file. Returns None if IOError
+
+        .. deprecated:: 2.0
+            no clear universal use case
 
         Returns
         -------
@@ -252,54 +179,30 @@ class Dsystem(object):
             widths = None
         return widths
 
-    def load_boundary_coord(self):
-        """
-        Load border coordinates from geometry path or from tokamak module.
-        Sets boundary_coord attribute to loaded coords.
-
-        Returns
-        -------
-        numpy.ndarray
-            contains (R, z) coordinates of vacuum vessel cross section points, should have shape (#points, 2)
-        """
-        warnings.warn('load_boundary_coord will be deprecated by load_border', FutureWarning)
-        bdr_path = self.geometry_path / 'border.txt'
-        boundary_coord = self.load_border(self, bdr_path)
-        if np.all(boundary_coord[:, 0] != boundary_coord[:, -1]):
-            boundary_coord = np.vstack([boundary_coord, boundary_coord[0, :][np.newaxis, :]])
-        return boundary_coord
-
-
-    def load_border(self, loc=None):
+    def load_border(self, shot: int = None, **kw) -> np.ndarray:
         """
         Load border coordinates from geometry path or from tokamak module.
         Sets boundary_coord attribute to loaded coords.
 
         Parameters
         ----------
-        loc : str or Path
+        shot : int, optional
+            shot number used to determine configuration, by default self.shot
 
         Returns
         -------
         numpy.ndarray
             contains (R, z) coordinates of vacuum vessel cross section, shape (#points, 2)
         """
-        if loc is None:
-            bdr_path = self.geometry_path / 'border.txt'
-        else:
-            bdr_path = Path(loc)
-        if not bdr_path.exists():
-            print('border coord not found at {}.'.format(bdr_path))
-            bdr_path = self.geometry_path.parent / 'border.txt'
-            print('Tyring to load from {}'.format(bdr_path))
-        bdr = np.loadtxt(str(bdr_path))
-        return bdr
+        raise NotImplementedError()
 
+# TODO: remove
     def compute_bd_mat(self, grid, loc=None):
         """
         Simple generation of boundary matrix 1 inside rec area, 0 outside
 
         .. deprecated:: 1.3.0
+            use :obj:`tomotok.core.grid.RegularGrid.is_inside` instead
 
         Parameters
         ----------